+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## In Progress
 
 ### Added
+- View state will automatically be set if it is not found for `Scatterplot` and `Spatial` components.
 
 ### Changed
 - Added a check for undefined in the `nodeToSet` function in `cell-set-utils.js`.
 - Fix bug which previously caused `anndata-cells.zarr` file types to fail to load when lacking a `"factors"` option.
+- `targetX` and `targetY` coordination values are `null` by default.
 
 ## [1.1.4](https://www.npmjs.com/package/vitessce/v/1.1.4) - 2021-02-11